+++ MailSender/View/RecipientsView.xaml
@@ -4,7 +4,6 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
-             xmlns:local="clr-namespace:WpfTestMailSender.View"
              mc:Ignorable="d" 
              DataContext="{Binding MainWindowModel,Source={StaticResource Locator}}"
              d:DesignHeight="450" d:DesignWidth="800">
