+++ MailSender/ViewModel/ViewModelLocator.cs
@@ -1,23 +1,7 @@
-/*
-  In App.xaml:
-  <Application.Resources>
-      <vm:ViewModelLocator xmlns:vm="clr-namespace:WpfTestMailSender"
-                           x:Key="Locator" />
-  </Application.Resources>
-  
-  In the View:
-  DataContext="{Binding Source={StaticResource Locator}, Path=ViewModelName}"
-
-  You can also use Blend to do all this with the tool's support.
-  See http://www.galasoft.ch/mvvm
-*/
-
-using GalaSoft.MvvmLight;
 using GalaSoft.MvvmLight.Ioc;
 using CommonServiceLocator;
 using MailSenderLib.Data.LinqToSql;
 using MailSenderLib.Services;
-using System;
 
 namespace WpfTestMailSender.ViewModel
 {
@@ -42,9 +26,9 @@ namespace WpfTestMailSender.ViewModel
                .TryRegister(() => new MailSenderDBDataContext());
 
             //services
-            //   .TryRegister<IRecipientsDataProvider, InMemoryRecipientsDataProvider>()
-            //   .TryRegister<ISendersDataProvider, InMemorySendersDataProvider>()
-            //   .TryRegister<IServersDataProvider, InMemoryServersDataProvider>();
+            //   .TryRegister<IRecipientsDataProvider, InMemoryMemoryRecipientsDataProvider>()
+            //   .TryRegister<ISendersDataProvider, InMemoryMemorySendersDataProvider>()
+            //   .TryRegister<IServersDataProvider, InMemoryMemoryServersDataProvider>();
         }
 
         public MainWindowViewModel MainWindowModel =>
@@ -55,37 +39,5 @@ namespace WpfTestMailSender.ViewModel
         {
             // TODO Clear the ViewModels
         }
-    }
-
-    public static class SimpleIocExtentions
-    {
-        public static SimpleIoc TryRegister<TInterface, TService>(this SimpleIoc services)
-            where TInterface : class
-            where TService : class, TInterface
-        {
-            if (services.IsRegistered<TInterface>()) 
-                return services;
-            services.Register<TInterface, TService>();
-            return services;
-        }
-
-        public static SimpleIoc TryRegister<TService>(this SimpleIoc services)
-           where TService : class
-        {
-            if (services.IsRegistered<TService>()) 
-                return services;
-            services.Register<TService>();
-            return services;
-        }
-
-        public static SimpleIoc TryRegister<TService>(this SimpleIoc services, Func<TService> Factory)
-           where TService : class
-        {
-            if (services.IsRegistered<TService>()) 
-                return services;
-            services.Register(Factory);
-            return services;
-        }
-
     }
 }
