+++ MailSenderLib/Services/InMemory/InMemoryRecipientsDataProvider.cs
@@ -6,7 +6,7 @@ using MailSenderLib.Entityes;
 
 namespace MailSenderLib.Services
 {
-    public class InMemoryRecipientsDataProvider: InDataProvider<Recipient>, IRecipientsDataProvider
+    public class InMemoryMemoryRecipientsDataProvider: InMemoryDataProvider<Recipient>, IRecipientsDataProvider
     {
         public override void Edit(int id, Recipient item)
         {
