+++ MailSenderLib/Services/InMemory/InMemoryRecipientsListsDataProvider.cs
@@ -2,7 +2,7 @@
 
 namespace MailSenderLib.Services
 {
-    public class InMemoryRecipientsListsDataProvider : InDataProvider<RecipientsList>, IRecipientsListsDataProvider
+    public class InMemoryMemoryRecipientsListsDataProvider : InMemoryDataProvider<RecipientsList>, IRecipientsListsDataProvider
     {
         public override void Edit(int id, RecipientsList item)
         {
